fix: clean up compiler warnings in project code
Remove unused variables (h, hp1, particle_ordinals_end), fix
signed/unsigned comparison mismatches, and add virtual destructor
to SnapshotWriter base class.

+++ core/pic/particle.h
@@ -320,7 +320,7 @@ template<std::ranges::forward_range R>
   static constexpr auto Nmax = 27uz;  // atm nothing should call this with more.
   const auto Nspans          = rn::distance(spans);
 
-  if(Nspans > Nmax) {
+  if(static_cast<std::size_t>(Nspans) > Nmax) {
     throw std::runtime_error {
       "ParticleContainer: trying to set from too many spans."
     };
